split the code into homescreen and saerchcomponent
